🌱 Split settings/Settings.tsx and settings/UpdateSettingsForm.tsx
Fixes #21796

Extract useSettingsTabs (view-model hook), Settings.parts (sidebar/header components), and UpdateSettingsForm.fields (field group components). Settings.tsx: 638→173 lines, 1 hook call. UpdateSettingsForm.tsx: 494→193 lines.
