+++ web/src/components/settings/Settings.tsx
@@ -1,46 +1,4 @@
-import { useEffect, useState, useRef, useCallback } from 'react'
-import { useTranslation } from 'react-i18next'
-import { useLocation, useNavigate } from 'react-router-dom'
-import {
-  Cpu,
-  TrendingUp,
-  Coins,
-  User,
-  Bell,
-  Shield,
-  Palette,
-  Eye,
-  Plug,
-  LayoutGrid,
-  Download,
-  Database,
-  Container,
-  HardDrive,
-  CheckCircle,
-  Loader2,
-  AlertCircle,
-  WifiOff,
-  BarChart3,
-} from 'lucide-react'
-import { Github } from '@/lib/icons'
-import { useAuth } from '../../lib/auth'
-import { useTheme } from '../../hooks/useTheme'
-import { useTokenUsage } from '../../hooks/useTokenUsage'
-import { useAIMode } from '../../hooks/useAIMode'
-import { useLocalAgent } from '../../hooks/useLocalAgent'
-import { useBackendHealth } from '../../hooks/useBackendHealth'
-import { useAccessibility } from '../../hooks/useAccessibility'
-import { useVersionCheck } from '../../hooks/useVersionCheck'
-import { usePredictionSettings } from '../../hooks/usePredictionSettings'
-import {
-  usePersistedSettings,
-  type SyncStatus,
-} from '../../hooks/usePersistedSettings'
-import {
-  BANNER_DISMISS_MS,
-  UI_FEEDBACK_TIMEOUT_MS,
-  TOOLTIP_HIDE_DELAY_MS,
-} from '../../lib/constants/network'
+import { useSettingsTabs } from './useSettingsTabs'
 import { UpdateSettings } from './UpdateSettings'
 import {
   AISettingsSection,
@@ -59,488 +17,73 @@ import {
   SettingsBackupSection,
   AnalyticsSection,
 } from './sections'
-import { cn } from '../../lib/cn'
-
-// Labels are filled at render time via t()
-const SYNC_ICONS: Record<
-  SyncStatus,
-  { icon: typeof CheckCircle; className: string }
-> = {
-  idle: { icon: CheckCircle, className: 'text-muted-foreground' },
-  saving: { icon: Loader2, className: 'text-yellow-400' },
-  saved: { icon: CheckCircle, className: 'text-green-400' },
-  error: { icon: AlertCircle, className: 'text-red-400' },
-  offline: { icon: WifiOff, className: 'text-muted-foreground' },
-}
-
-// Define settings navigation structure with groups
-// Labels use i18n keys resolved at render time
-const SETTINGS_NAV = [
-  {
-    groupKey: 'settings.groups.aiIntelligence' as const,
-    items: [
-      {
-        id: 'ai-mode-settings',
-        labelKey: 'settings.nav.aiMode' as const,
-        icon: Cpu,
-      },
-      {
-        id: 'prediction-settings',
-        labelKey: 'settings.nav.predictions' as const,
-        icon: TrendingUp,
-      },
-      {
-        id: 'agent-settings',
-        labelKey: 'settings.nav.localAgent' as const,
-        icon: Plug,
-      },
-      {
-        id: 'token-usage-settings',
-        labelKey: 'settings.nav.tokenUsage' as const,
-        icon: Coins,
-      },
-    ],
-  },
-  {
-    groupKey: 'settings.groups.integrations' as const,
-    items: [
-      {
-        id: 'github-token-settings',
-        labelKey: 'settings.nav.github' as const,
-        icon: Github,
-      },
-      {
-        id: 'widget-settings',
-        labelKey: 'settings.nav.desktopWidget' as const,
-        icon: LayoutGrid,
-      },
-      {
-        id: 'persistence-settings',
-        labelKey: 'settings.nav.deployPersistence' as const,
-        icon: Database,
-      },
-    ],
-  },
-  {
-    groupKey: 'settings.groups.userAlerts' as const,
-    items: [
-      {
-        id: 'profile-settings',
-        labelKey: 'settings.nav.profile' as const,
-        icon: User,
-      },
-      {
-        id: 'notifications-settings',
-        labelKey: 'settings.nav.notifications' as const,
-        icon: Bell,
-      },
-    ],
-  },
-  {
-    groupKey: 'settings.groups.appearance' as const,
-    items: [
-      {
-        id: 'theme-settings',
-        labelKey: 'settings.nav.theme' as const,
-        icon: Palette,
-      },
-      {
-        id: 'accessibility-settings',
-        labelKey: 'settings.nav.accessibility' as const,
-        icon: Eye,
-      },
-    ],
-  },
-  {
-    groupKey: 'settings.groups.utilities' as const,
-    items: [
-      {
-        id: 'settings-backup',
-        labelKey: 'settings.nav.backupSync' as const,
-        icon: HardDrive,
-      },
-      {
-        id: 'local-clusters-settings',
-        labelKey: 'settings.nav.localClusters' as const,
-        icon: Container,
-      },
-      {
-        id: 'permissions-settings',
-        labelKey: 'settings.nav.permissions' as const,
-        icon: Shield,
-      },
-      {
-        id: 'analytics-settings',
-        labelKey: 'settings.nav.analytics' as const,
-        icon: BarChart3,
-      },
-      {
-        id: 'system-updates-settings',
-        labelKey: 'settings.nav.updates' as const,
-        icon: Download,
-      },
-    ],
-  },
-]
+import {
+  SettingsSidebar,
+  SectionGroupHeader,
+  RestoredToast,
+  MobileHeader,
+} from './Settings.parts'
 
 export function Settings() {
-  const { t } = useTranslation()
-  const location = useLocation()
-  const navigate = useNavigate()
-  const { user, refreshUser, isLoading: isUserLoading } = useAuth()
-  const { themeId, setTheme, themes, currentTheme } = useTheme()
-  const { usage, updateSettings, resetUsage, isDemoData } = useTokenUsage()
-  const { mode, setMode, description } = useAIMode()
-  const { health, isConnected, refresh } = useLocalAgent()
-  const { isInClusterMode } = useBackendHealth()
   const {
+    activeSection,
+    showRestoredToast,
+    contentRef,
+    handleNavClick,
+    user,
+    refreshUser,
+    isUserLoading,
+    themeId,
+    setTheme,
+    themes,
+    currentTheme,
+    usage,
+    updateTokenSettings,
+    resetUsage,
+    isDemoData,
+    mode,
+    setMode,
+    description,
+    health,
+    isConnected,
+    refreshAgent,
+    isInClusterMode,
     colorBlindMode,
     setColorBlindMode,
     reduceMotion,
     setReduceMotion,
     highContrast,
     setHighContrast,
-  } = useAccessibility()
-  const { forceCheck: forceVersionCheck } = useVersionCheck()
-  const {
-    settings: predictionSettings,
-    updateSettings: updatePredictionSettings,
-    resetSettings: resetPredictionSettings,
-  } = usePredictionSettings()
-  const {
-    restoredFromFile,
+    forceVersionCheck,
+    predictionSettings,
+    updatePredictionSettings,
+    resetPredictionSettings,
     syncStatus,
     lastSaved,
     filePath,
     exportSettings,
     importSettings,
-  } = usePersistedSettings()
-
-  const [activeSection, setActiveSection] = useState<string>('ai-mode-settings')
-  const [showRestoredToast, setShowRestoredToast] = useState(false)
-
-  // Suppresses IntersectionObserver updates during programmatic scrolls
-  // so the sidebar highlight stays on the clicked item instead of flickering
-  // through intermediate sections while the smooth scroll animates.
-  const isNavScrollingRef = useRef(false)
-  const navScrollTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null)
-
-  /** Duration to suppress observer after a nav click (covers smooth scroll animation).
-   *  Long scrolls (e.g. AI Mode → Updates) can exceed 800ms on some browsers. */
-  const NAV_SCROLL_SUPPRESS_MS = 1200
-
-  // Show toast when settings are restored from backup file (after cache clear)
-  // Note: settings uses sync status (lastSaved) for persistence tracking, not API data timestamps
-  useEffect(() => {
-    if (restoredFromFile) {
-      setShowRestoredToast(true)
-      const timer = setTimeout(
-        () => setShowRestoredToast(false),
-        BANNER_DISMISS_MS,
-      )
-      return () => clearTimeout(timer)
-    }
-  }, [restoredFromFile])
-  const contentRef = useRef<HTMLDivElement>(null)
-
-  // Offset so scrolled-to sections land with breathing room (accounts for demo banner + visual centering)
-  const SCROLL_OFFSET = 80
-
-  const getScrollContainer = () => document.getElementById('main-content')
-
-  const scrollToSection = useCallback((sectionId: string, smooth = true) => {
-    const element = document.getElementById(sectionId)
-    const container = getScrollContainer()
-    if (!element || !container) return
-    const containerRect = container.getBoundingClientRect()
-    const elementRect = element.getBoundingClientRect()
-    const y =
-      elementRect.top - containerRect.top + container.scrollTop - SCROLL_OFFSET
-    container.scrollTo({ top: y, behavior: smooth ? 'smooth' : 'auto' })
-  }, [])
-
-  // Handle deep linking — scroll to section based on URL hash.
-  // Depends on both pathname and hash so it fires when navigating TO settings
-  // from another page (KeepAlive keeps Settings mounted, so location updates
-  // for all routes — we only act when actually on /settings).
-  // Skipped when isNavScrollingRef is set (handleNavClick already scrolled).
-  useEffect(() => {
-    if (location.pathname !== '/settings') return
-    if (isNavScrollingRef.current) return
-    const hash = location.hash.replace('#', '')
-    if (!hash) return
-
-    // Retry scroll a few times — KeepAlive transitions display:none→contents
-    // and the element may not have a layout rect on the first frame.
-    let attempts = 0
-    const maxAttempts = 5
-    const tryScroll = () => {
-      const element = document.getElementById(hash)
-      const container = getScrollContainer()
-      if (element && container && element.getBoundingClientRect().height > 0) {
-        // Suppress observer while deep-link scroll settles
-        isNavScrollingRef.current = true
-        if (navScrollTimerRef.current) clearTimeout(navScrollTimerRef.current)
-        navScrollTimerRef.current = setTimeout(() => {
-          isNavScrollingRef.current = false
-        }, NAV_SCROLL_SUPPRESS_MS)
-
-        scrollToSection(hash, false)
-        setActiveSection(hash)
-        element.classList.add('ring-2', 'ring-purple-500/50')
-        setTimeout(
-          () => element.classList.remove('ring-2', 'ring-purple-500/50'),
-          UI_FEEDBACK_TIMEOUT_MS,
-        )
-      } else if (++attempts < maxAttempts) {
-        requestAnimationFrame(tryScroll)
-      }
-    }
-    // Initial delay for route transition, then retry with rAF
-    const timer = setTimeout(tryScroll, TOOLTIP_HIDE_DELAY_MS)
-    return () => clearTimeout(timer)
-  }, [location.pathname, location.hash, scrollToSection])
-
-  // Track active section on scroll using IntersectionObserver.
-  // Debounced to prevent a feedback loop: observer fires -> activeSection
-  // changes -> sidebar re-renders -> sticky layout shifts -> observer fires
-  // again with a different section -> rapid flickering. The debounce
-  // coalesces rapid-fire observer callbacks into a single state update.
-  /** Debounce interval for observer-driven activeSection updates (ms). */
-  const OBSERVER_DEBOUNCE_MS = 100
-  useEffect(() => {
-    const container = getScrollContainer()
-    if (!container) return
-
-    const allSectionIds = SETTINGS_NAV.flatMap((g) => g.items.map((i) => i.id))
-    const visibleSections = new Map<string, number>()
-    let debounceTimer: ReturnType<typeof setTimeout> | null = null
-
-    const observer = new IntersectionObserver(
-      (entries) => {
-        // Always keep visibleSections accurate so stale data doesn't
-        // cause wrong highlights after programmatic scrolls end.
-        for (const entry of entries) {
-          if (entry.isIntersecting) {
-            visibleSections.set(entry.target.id, entry.intersectionRatio)
-          } else {
-            visibleSections.delete(entry.target.id)
-          }
-        }
-        // Skip activeSection updates during programmatic scrolls
-        // (nav click or deep link) — handleNavClick already set it.
-        if (isNavScrollingRef.current) return
-
-        // Debounce to prevent scroll-position feedback loops:
-        // observer -> setState -> sidebar re-render -> layout shift -> observer
-        if (debounceTimer) clearTimeout(debounceTimer)
-        debounceTimer = setTimeout(() => {
-          // Pick the first visible section in document order
-          for (const id of allSectionIds) {
-            if (visibleSections.has(id)) {
-              // Functional update avoids unnecessary re-renders when
-              // the computed section hasn't actually changed.
-              setActiveSection((prev) => prev === id ? prev : id)
-              break
-            }
-          }
-        }, OBSERVER_DEBOUNCE_MS)
-      },
-      {
-        root: container,
-        // Use a single threshold to reduce observer callback frequency.
-        // Multiple thresholds (e.g. [0, 0.1, 0.5]) fire at each boundary
-        // crossing, amplifying the feedback loop.
-        rootMargin: '0px 0px -40% 0px',
-        threshold: 0 }
-    )
-
-    for (const id of allSectionIds) {
-      const el = document.getElementById(id)
-      if (el) observer.observe(el)
-    }
-
-    return () => {
-      observer.disconnect()
-      if (debounceTimer) clearTimeout(debounceTimer)
-    }
-  }, [])
-
-  const handleNavClick = (sectionId: string) => {
-    // Suppress IntersectionObserver while the smooth scroll animates
-    isNavScrollingRef.current = true
-    if (navScrollTimerRef.current) clearTimeout(navScrollTimerRef.current)
-    navScrollTimerRef.current = setTimeout(() => {
-      isNavScrollingRef.current = false
-    }, NAV_SCROLL_SUPPRESS_MS)
-
-    scrollToSection(sectionId)
-    setActiveSection(sectionId)
-    // Update URL hash without triggering the deep link effect (isNavScrollingRef guards it)
-    navigate(`#${sectionId}`, { replace: true })
-
-    // Keep the clicked item visible in the sidebar's own scroll area.
-    // NOTE: scrollIntoView cascades to ALL ancestor scroll containers
-    // (including #main-content), which cancels the smooth scroll set by
-    // scrollToSection above. Instead, manually scroll only the sidebar's
-    // own overflow container so #main-content is unaffected.
-    requestAnimationFrame(() => {
-      const btn = document.querySelector<HTMLElement>(
-        `[data-settings-nav="${sectionId}"]`,
-      )
-      if (!btn) return
-      const sidebar = btn.closest<HTMLElement>('.overflow-y-auto')
-      if (!sidebar || sidebar.id === 'main-content') return
-      const btnRect = btn.getBoundingClientRect()
-      const sidebarRect = sidebar.getBoundingClientRect()
-      if (
-        btnRect.top < sidebarRect.top ||
-        btnRect.bottom > sidebarRect.bottom
-      ) {
-        const scrollDelta =
-          btnRect.top -
-          sidebarRect.top -
-          sidebarRect.height / 2 +
-          btnRect.height / 2
-        sidebar.scrollBy({ top: scrollDelta, behavior: 'smooth' })
-      }
-    })
-  }
-
-  const SYNC_LABELS: Record<SyncStatus, string> = {
-    idle: t('settings.syncStatus.synced'),
-    saving: t('settings.syncStatus.saving'),
-    saved: t('settings.syncStatus.savedToFile'),
-    error: t('settings.syncStatus.saveFailed'),
-    offline: t('settings.syncStatus.localOnly'),
-  }
-  const sync = SYNC_ICONS[syncStatus]
-  const SyncIcon = sync.icon
-  const syncLabel = SYNC_LABELS[syncStatus]
+  } = useSettingsTabs()
 
   return (
     <div
       data-testid="settings-page"
       className="pt-16 max-w-6xl mx-auto flex flex-col lg:flex-row gap-6"
     >
-      {/* Settings restored toast */}
-      {showRestoredToast && (
-        <div className="fixed top-20 right-4 z-toast bg-green-500/20 border border-green-500/30 text-green-400 px-4 py-2 rounded-lg text-sm shadow-lg backdrop-blur-xs animate-in slide-in-from-right">
-          {t('settings.restoredFromBackup')}
-        </div>
-      )}
-      {/* Sidebar Navigation */}
-      <nav className="hidden lg:block w-56 shrink-0">
-        <div className="sticky top-20 max-h-[calc(100vh-6rem)] overflow-y-auto scroll-enhanced space-y-4">
-          <div className="mb-4">
-            <div className="flex items-center justify-between">
-              <h1
-                data-testid="settings-title"
-                className="text-xl font-bold text-foreground"
-                data-qa="settings-header"
-              >
-                {t('settings.title')}
-              </h1>
-            </div>
-            <p className="text-sm text-muted-foreground">
-              {t('settings.subtitle')}
-            </p>
-            <div
-              className={cn(
-                'flex items-center gap-1.5 mt-2 text-xs',
-                sync.className,
-              )}
-            >
-              <SyncIcon
-                className={cn(
-                  'w-3.5 h-3.5',
-                  syncStatus === 'saving' && 'animate-spin',
-                )}
-              />
-              <span>{syncLabel}</span>
-            </div>
-          </div>
-          {SETTINGS_NAV.map((group) => (
-            <div key={group.groupKey}>
-              <h3 className="text-2xs uppercase tracking-wider text-muted-foreground font-semibold mb-1 px-2">
-                {t(group.groupKey)}
-              </h3>
-              <div className="space-y-0.5">
-                {group.items.map((item) => {
-                  const Icon = item.icon
-                  const isActive = activeSection === item.id
-                  return (
-                    <button
-                      key={item.id}
-                      data-settings-nav={item.id}
-                      onClick={() => handleNavClick(item.id)}
-                      aria-label={t('actions.openSettingsSectionAria', { section: t(item.labelKey) })}
-                      className={cn(
-                        'w-full flex items-center gap-2 px-2 py-1.5 rounded-md text-sm transition-colors text-left',
-                        isActive
-                          ? 'bg-purple-500/20 text-purple-400'
-                          : 'text-muted-foreground hover:text-foreground hover:bg-secondary/50',
-                      )}
-                    >
-                      <Icon
-                        className={cn(
-                          'w-4 h-4 shrink-0',
-                          isActive
-                            ? 'text-purple-400'
-                            : 'text-muted-foreground',
-                        )}
-                      />
-                      <span className="truncate">{t(item.labelKey)}</span>
-                    </button>
-                  )
-                })}
-              </div>
-            </div>
-          ))}
-        </div>
-      </nav>
-
-      {/* Main Content */}
+      <RestoredToast show={showRestoredToast} />
+      <SettingsSidebar
+        activeSection={activeSection}
+        onNavClick={handleNavClick}
+        syncStatus={syncStatus}
+      />
       <div ref={contentRef} className="flex-1 min-w-0">
-        {/* Mobile Header */}
-        <div className="block lg:hidden mb-6">
-          <div className="flex items-center justify-between">
-            <h1
-              data-testid="settings-title"
-              className="text-2xl font-bold text-foreground"
-              data-qa="settings-header-mobile"
-            >
-              {t('settings.title')}
-            </h1>
-          </div>
-          <p className="text-muted-foreground">{t('settings.subtitle')}</p>
-          <div
-            className={cn(
-              'flex items-center gap-1.5 mt-2 text-xs',
-              sync.className,
-            )}
-          >
-            <SyncIcon
-              className={cn(
-                'w-3.5 h-3.5',
-                syncStatus === 'saving' && 'animate-spin',
-              )}
-            />
-            <span>{syncLabel}</span>
-          </div>
-        </div>
+        <MobileHeader syncStatus={syncStatus} />
 
         {/* AI & Intelligence Group */}
         <div className="mb-8">
-          <h2 className="text-xs uppercase tracking-wider text-muted-foreground font-semibold mb-3 px-1">
-            {t('settings.groups.aiIntelligence')}
-          </h2>
+          <SectionGroupHeader labelKey="settings.groups.aiIntelligence" />
           <div className="space-y-6">
-            <AISettingsSection
-              mode={mode}
-              setMode={setMode}
-              description={description}
-            />
+            <AISettingsSection mode={mode} setMode={setMode} description={description} />
             <PredictionSettingsSection
               settings={predictionSettings}
               updateSettings={updatePredictionSettings}
@@ -550,11 +93,11 @@ export function Settings() {
               isConnected={isConnected}
               isInClusterMode={isInClusterMode}
               health={health}
-              refresh={refresh}
+              refresh={refreshAgent}
             />
             <TokenUsageSection
               usage={usage}
-              updateSettings={updateSettings}
+              updateSettings={updateTokenSettings}
               resetUsage={resetUsage}
               isDemoData={isDemoData}
             />
@@ -563,9 +106,7 @@ export function Settings() {
 
         {/* Integrations Group */}
         <div className="mb-8">
-          <h2 className="text-xs uppercase tracking-wider text-muted-foreground font-semibold mb-3 px-1">
-            {t('settings.groups.integrations')}
-          </h2>
+          <SectionGroupHeader labelKey="settings.groups.integrations" />
           <div className="space-y-6">
             <GitHubTokenSection forceVersionCheck={forceVersionCheck} />
             <WidgetSettingsSection />
@@ -575,9 +116,7 @@ export function Settings() {
 
         {/* User & Alerts Group */}
         <div className="mb-8">
-          <h2 className="text-xs uppercase tracking-wider text-muted-foreground font-semibold mb-3 px-1">
-            {t('settings.groups.userAlerts')}
-          </h2>
+          <SectionGroupHeader labelKey="settings.groups.userAlerts" />
           <div className="space-y-6">
             <ProfileSection
               initialEmail={user?.email ?? ''}
@@ -592,9 +131,7 @@ export function Settings() {
 
         {/* Appearance Group */}
         <div className="mb-8">
-          <h2 className="text-xs uppercase tracking-wider text-muted-foreground font-semibold mb-3 px-1">
-            {t('settings.groups.appearance')}
-          </h2>
+          <SectionGroupHeader labelKey="settings.groups.appearance" />
           <div className="space-y-6">
             <ThemeSection
               themeId={themeId}
@@ -615,9 +152,7 @@ export function Settings() {
 
         {/* Utilities Group */}
         <div className="mb-8">
-          <h2 className="text-xs uppercase tracking-wider text-muted-foreground font-semibold mb-3 px-1">
-            {t('settings.groups.utilities')}
-          </h2>
+          <SectionGroupHeader labelKey="settings.groups.utilities" />
           <div className="space-y-6">
             <SettingsBackupSection
               syncStatus={syncStatus}
